Move key storage config to dedicated location for improved documentation
